Start adding unit tests
