+++ opsys/opsys.go
@@ -65,8 +65,8 @@ type (
 		// path is returned.
 		HEIC2JPG(src string) (string, error)
 		NewTxtOutFile(afero.File) OutFile
-		NewWeasyPrintOutFile(chatFile afero.File, includePPA bool) OutFile
-		NewWkhtmltopdfOutFile(chatFile afero.File, pdfg pdfgen.PDFGenerator, includePPA bool) OutFile
+		NewWeasyPrintFile(chatFile afero.File, includePPA bool) OutFile
+		NewWkhtmltopdfFile(chatFile afero.File, pdfg pdfgen.PDFGenerator, includePPA bool) OutFile
 	}
 
 	opSys struct {
